feat: enhance web search functionality with improved configuration and environment variable extraction for authFields
