+++ packages/data-provider/src/config.ts
@@ -1310,8 +1310,6 @@ export enum Constants {
   mcp_delimiter = '_mcp_',
   /** Placeholder Agent ID for Ephemeral Agents */
   EPHEMERAL_AGENT_ID = 'ephemeral',
-  /** Prefix for LC Dynamic Env. Variables */
-  LC_VAR_PREFIX = 'LC_VAR_',
 }
 
 export enum LocalStorageKeys {
