+++ packages/data-provider/src/index.ts
@@ -27,6 +27,7 @@ export * from './types/files';
 export * from './types/mutations';
 export * from './types/queries';
 export * from './types/runs';
+export * from './types/web';
 /* query/mutation keys */
 export * from './keys';
 /* api call helpers */
